Remove unused code YQL-17437
