--- ydb/library/yql/core/cbo/cbo_optimizer.h
@@ -1,75 +1,0 @@
-#pragma once
-
-#include <util/generic/vector.h>
-
-#include <unordered_map>
-#include <map>
-
-namespace NYql {
-
-struct IOptimizer {
-    using TVarId = std::tuple<int, int>;
-
-    struct TVar {
-        char Name = 0; // debug name: 'a', 'b', 'c', ...
-    };
-
-    struct TRel {
-        double Rows = 0;
-        double TotalCost = 0;
-
-        std::vector<TVar> TargetVars;
-    };
-
-    struct TEq {
-        std::vector<TVarId> Vars;
-    };
-
-    struct TInput {
-        std::vector<TRel> Rels;
-        std::vector<TEq> EqClasses;
-        std::vector<TEq> Left;
-        std::vector<TEq> Right;
-
-        TString ToString() const;
-        void Normalize();
-    };
-
-    enum class EJoinType {
-        Unknown,
-        Inner,
-        Left,
-        Right,
-    };
-
-    enum class EJoinStrategy {
-        Unknown,
-        Hash,
-        Loop
-    };
-
-    struct TJoinNode {
-        EJoinType Mode = EJoinType::Unknown;
-        EJoinStrategy Strategy = EJoinStrategy::Unknown;
-        // only a = b && c = d ...  supported yet
-        std::vector<TVarId> LeftVars = {};
-        std::vector<TVarId> RightVars = {};
-        std::vector<int> Rels = {};
-        int Outer = -1; // index in Nodes
-        int Inner = -1; // index in Nodes
-    };
-
-    struct TOutput {
-        std::vector<TJoinNode> Nodes;
-        TInput* Input = nullptr;
-        double Rows = 0;
-        double TotalCost = 0;
-
-        TString ToString(bool printCost = true) const;
-    };
-
-    virtual ~IOptimizer() = default;
-    virtual TOutput JoinSearch() = 0;
-};
-
-} // namespace NYql
