+++ ydb/library/yql/core/cbo/ya.make
@@ -1,12 +1,7 @@
 LIBRARY()
 
 SRCS(
-    cbo_optimizer.cpp
     cbo_optimizer_new.cpp
-)
-
-PEERDIR(
-    library/cpp/disjoint_sets
 )
 
 END()
